requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
colorlog==4.8.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
filetype==1.0.13
har2case==0.3.1
HttpRunner==1.5.8
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonpath==0.82
jsonschema==3.2.0
MarkupSafe==2.0.1
packaging==21.3
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.0.8
pyrsistent==0.18.1
pytz==2022.1
PyYAML==5.4.1
requests==2.27.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==0.13.5
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
view==0.1
